Directory refactor for better classification of classes
