+++ src/classModel/Catalog.ts
