+++ src/classModel/Product.ts
